requirements.txt
typer[all]
rich
pathlib
pydantic
beautifulsoup4
lxml
openai>=1.2.0  # For Azure OpenAI SDK
azure-identity>=1.12.0
azure-core>=1.26.0
python-dotenv
tavily-python